+++ src/main/java/com/develcode/develfood/services/exception/validation/CnpjExistsException.java
@@ -1,4 +1,4 @@
-package com.develcode.develfood.services.exception;
+package com.develcode.develfood.services.exception.validation;
 
 public class CnpjExistsException extends RuntimeException{
     public CnpjExistsException(String message) {super(message);}
