+++ src/main/java/com/develcode/develfood/services/exception/validation/CpfExistsException.java
@@ -1,4 +1,4 @@
-package com.develcode.develfood.services.exception;
+package com.develcode.develfood.services.exception.validation;
 
 public class CpfExistsException extends RuntimeException{
     public CpfExistsException(String message) {super(message);}
