+++ src/main/java/com/develcode/develfood/services/exception/validation/EmailExistsException.java
@@ -1,4 +1,4 @@
-package com.develcode.develfood.services.exception;
+package com.develcode.develfood.services.exception.validation;
 
 public class EmailExistsException extends RuntimeException{
     public EmailExistsException(String message) {super(message);}
