+++ src/main/java/com/develcode/develfood/services/exception/validation/ValidationErrorData.java
@@ -1,0 +1,5 @@
+package com.develcode.develfood.services.exception.validation;
+
+public class ValidationErrorData extends RuntimeException{
+
+}
